Include is raw code to implement a .NET TVMClient. This is include per discussion with Wade Matveyenko.
This code does not compile and as such the project file has not been updated. This code was originally written to support a Windows Phone based TVMClient and has a dependency on BouncyCastle.Crypto.WP7. Additionally, it uses Action Delegates with multiple type parameters which wasn't introduced until .NET 3.5 and the current AWSSDK targets 2.0.

Where available, code calls into the AWSSDKUtil for helpers.