Avoid conflicts that would occur between serial updates to the record by the same device during sync.
When a record is updated during the sync operation, it's meta data will not be updated, even though the prior local modification was written to the server, and the sync count was updated.  During the next sync operation, the put to remote will fail, due to a data conflict exception (modifying a stale sync count).  The data conflict exception handling will then fall back to an older sync count for the dataset and try again.  This will lead to a case where the local change (with an older sync count) conflicts with a remote change (older, but with a more recent sync count).

To avoid this conflict, update the local records meta data during the first push.  This means that on a subsequent push, there will not be a data conflict exception, and ultimately there will not be a conflict-with-self where the sync counts reflect the opposite of what is actually true in terms of data recency.

The goal is to make it look like the put operation succeeded, wrote the remote record to local storage, and *then* performed the update (instead of performing the update before the put send/receive had a chance to finish).