Make the local merge dataset sync count match the remote sync count
Once the merge data set is created under the new authenticated cognito id, the sync count is 1 (even though records in that data set are unmodified and probably higher than 1).  Since the local merge dataset does not have a modified sync count, when a synchronize is performed prior to a deletion, the synchronize fails, with a message that the local sync count does not exist.  This modification updates the sync count in the local data store to 1, so that it matches the one on the server.