Previous implementation of retrieving credentials from EC2 IMDS would hammer the service when too many clients were being created in short amount of time. We decided to implement a singleton class to address that problem. The change introduced a regression where too many client creation would create a thread pool bottle-neck.
The solution here is to create service clients without the initial fetch
of the credentials--it would just start the timer task and return the
instance immediately yielding the thread to other tasks.

When the DefaultInstanceProfileAWSCredentials.GetCredentials() is
called, we would hit IMDS if no credential has been cached already.  The
perf is actually not that bad, since the GetStringContentAsync() gets to
execute, in a stress test, we only hit IMDS 2~3 times before the timer
task runs and cached credentials is available.