Update AmazonSecurityTokenServiceConfig.cs
the call to FallbackRegionFactory.GetRegionEndpoint(false) throws exceptions when the configuration is not set. This causes problems when the application has no configuration, does not intend to ever have one, and AmazonSecurityTokenServiceClient is instantiated by passing the RegionEndpoint in the constructor. It really should never call FallbackRegionFactory.GetRegionEndpoint(false) if the RegionEndpoint is passed to the constructor. This caused me an entire day of frustration, and required me to set AWSConfig.Region in my application startup.