fix: sign AWSSDK.Core and extensions DLLs
## Motivation

Currently AWSSDK.Core shipped in the zip is not signed and none of the extensions dlls are signed.

## Changes

Signing step was working right after SDK dlls are built which excludes the extensions dlls from the signing. This change make sure that signing happens after extensions are built.

Secondly, include AWSSDK.Extensions*.dll to the signable items.

Some logging improvements which helped debugging the issue.

## Result

All AWSSDK.*.dll are now signed.