AWS SDK Compatibility: Check for `http_proxy` & `https_proxy` environment variables (#2991)
* check for `http_proxy` & `https_proxy` env-vars

this brings us in line with many other AWS SDKs which check for
these environment variables to modify client behavior. it does not
change behavior if you've already been setting your web proxy
explicitly, or in properties (the same as the java (and other) sdks).

we do introduce this with two new methods in ClientConfig to not break
any existing callers using the methods that are public. while all of our
callers have been updated to properly check web proxy, and then if
that's unset check the protocol being used and optionally call
`GetHttpProxy`/`GetHttpsProxy` as needed. If we're okay with breaking
ever IT MAY be worthwhile to break `GetWebProxy` by taking in the
protocol, and then just having the one method to prevent any accidental
mis-use, and to ensure everyone is using those environment variables.

it's also important to note:
<https://github.com/dotnet/runtime/issues/31113> meaning a user will
just get an error if they try to proxy to something that is using
`HTTPS`. i figured just letting it error on construction is probably
the 'safest' option. letting the user know that their setting isn't
being respected, as opposed to just silently 'ignoring' it. this is
also a decision we might want to change.

* actually parse out credentials for proxy

* remove accidental .swp file

* swap reading explicit over env-var first, tests

Ensure we read explicitly configured proxies through code first,
and then fallback to reading the proxy from environment variables.
next add some tests with the recommended environment variable mocking
test setup provided.

* address comments about BCL ClientConfig