Remove CreateAdditionalModelRequestFields from BedrockChatClient
Unknown parameters cause models to hard fail, and most of these settings are not supported by most of the models exposed in bedrock. A developer can use RawRepresentationFactory to populate these themselves if they really want them, so for now at least, this just gets rid of the possibly problematic settings.